Refactor test suite
For some reason there were way more fixtures then there were suppose to be. This commit refactors the fixtures so they're included in the only classes they're used and reduces them to only the minimum amount needed.
