--- tests/Fixtures/AlgoliaEngineTestModel.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Laravel\Scout\Tests\Fixtures;
-
-class AlgoliaEngineTestModel extends TestModel
-{
-    //
-}
