--- tests/Fixtures/EmptyTestModel.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Laravel\Scout\Tests\Fixtures;
-
-class EmptyTestModel extends TestModel
-{
-    public function toSearchableArray()
-    {
-        return [];
-    }
-}
