--- tests/Fixtures/SearchableTestModel.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Laravel\Scout\Tests\Fixtures;
-
-use Laravel\Scout\Searchable;
-
-class SearchableTestModel extends TestModel
-{
-    use Searchable;
-}
